+++ example/src/babel.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  presets: ["@babel/react"],
+  plugins: [["babel-plugin-styled-components", "transform-runtime"]]
+};