+++ src/libs/elements/Image/Styled.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { width, height } from "styled-system";
+
+export const ImageElement = styled.img`
+  ${width};
+  ${height};
+`;
+ImageElement.displayName = "ImageElement";