+++ CHANGELOG.md
@@ -14,8 +14,7 @@ Add:
 - `align-self: center` to `AvatarContent` of `AvatarGroup`
 - set `fluid` prop for `TextArea`
 - `Typo`: fix heading typo consist with origin heading tag, add `margin, fontWeight, fontFamily` props
-
-* image fluid prop
+- `Image`: add `width, height` props
 
 * `PushMessage`
 * `List`